Refactor to modular package structure, moving functionality to analytics/ subpackage
