+++ pyproject.toml
@@ -4,18 +4,18 @@ version = "0.1.0"
 description = "Add your description here"
 requires-python = ">=3.11"
 dependencies = [
-    "email-validator>=2.2.0",
+    # "email-validator>=2.2.0", # Removed - not used
     "flask>=3.1.0",
     "flask-sqlalchemy>=3.1.1",
     "flask-wtf>=1.2.0",
     "gunicorn>=23.0.0",
-    "numpy>=2.2.4",
+    "numpy>=2.2.4", # Used by pandas
     "pandas>=2.2.3",
-    "plotly>=6.0.1",
-    "psycopg2-binary>=2.9.10",
+    "plotly>=6.0.1", # Used for charts (even if legacy)
+    "psycopg2-binary>=2.9.10", # For optional PostgreSQL
     "yfinance>=0.2.55",
     "Flask-Migrate>=4.0.5",
-    "matplotlib>=3.8.0",
+    # "matplotlib>=3.8.0", # Removed - not used in provided files
 ]
 
 [project.optional-dependencies]
